requirements.txt
dify_plugin>=0.2.0
python-office>=0.5.0
pytest>=7.0.0